Pull Request 1
This pull request enhances the representation and handling of different tiles in the game. Currently, the game uses simple numeric values to represent different types of tiles. This update introduces a more robust system using objects and classes to represent tiles, making the code more maintainable and scalable. The new Tile class will allow for easy addition of new tile types in the future, and the GameMap class will manage the creation and display of the game map. Please review the new tile representation system. Any suggestions for improving the scalability or maintainability of this approach are welcome.